+++ lib/service/groups.dart
@@ -127,7 +127,6 @@ class Groups with Service implements NotificationsListener {
 
     _userGroups = await _loadUserGroupsFromCache();
     _userGroupNames = _buildGroupNames(_userGroups);
-    _initUserGroupsWithCachedAttendedMembers();
 
     if (_userGroups == null) {
       await _initUserGroupsFromNet();
@@ -226,6 +225,9 @@ class Groups with Service implements NotificationsListener {
 
   // Groups APIs
 
+  // MD: This method is important for user data migration.
+  // This method is supposed to run only once in the app lifecycle with user login due to the  _loggedIn flag
+  // Please keep in mind there may be old legacy users that haven't been migrated yet and this method resolves this scenario. 
   @protected
   Future<void> waitForLogin() async{
     if(!_loggedIn && Auth2().isLoggedIn) {
@@ -270,25 +272,55 @@ class Groups with Service implements NotificationsListener {
       }
   }
 
-  Future<List<Group>?> loadGroups({GroupsContentType? contentType}) async {
+  ///
+  /// Do not load user groups on portions / pages. We cached and use them for checks in flexUi and checklist
+  ///
+  /// Note: Do not allow loading on portions (paging) - there is a problem on the backend. Revert when it is fixed. 
+  Future<List<Group>?> loadGroups({GroupsContentType? contentType, String? category}) async {
     if (contentType == GroupsContentType.my) {
       await _waitForUpdateUserGroupsFromNet();
       return userGroups;
     } else {
-      return await _loadAllGroups();
+      return await _loadAllGroups(category: category);
     }
   }
 
-  Future<List<Group>?> _loadAllGroups() async {
+  Future<List<Group>?> _loadAllGroups({String? category, String? title, GroupPrivacy? privacy}) async {
     await waitForLogin();
     if (Config().groupsUrl != null) {
+      Map<String, String> queryParams = {};
+      if (StringUtils.isNotEmpty(category)) {
+        queryParams.addAll({'category': category!});
+      }
+      if (StringUtils.isNotEmpty(title)) {
+        queryParams.addAll({'title': title!});
+      }
+      if (privacy != null) {
+        queryParams.addAll({'privacy': groupPrivacyToString(privacy)!});
+      }
+      /*
+      // TMP disable paging - there is a problem on the backend
+      if (offset != null) {
+        queryParams.addAll({'offset': offset.toString()});
+      }
+      if (limit != null) {
+        queryParams.addAll({'limit': limit.toString()});
+      }*/
+      String url = '${Config().groupsUrl}/v2/groups';
+      if (queryParams.isNotEmpty) {
+        url = UrlUtils.addQueryParameters(url, queryParams);
+      }
+      
       try {
-        String url = '${Config().groupsUrl}/groups';
         Response? response = await Network().get(url, auth: Auth2());
         int responseCode = response?.statusCode ?? -1;
         String? responseBody = response?.body;
-        List<dynamic>? groupsJson = ((responseBody != null) && (responseCode == 200)) ? JsonUtils.decodeList(responseBody) : null;
-        return (groupsJson != null) ? Group.listFromJson(groupsJson) : null;
+        if (responseCode == 200) {
+          List<dynamic>? groupsJson = JsonUtils.decodeList(responseBody);
+          return Group.listFromJson(groupsJson);
+        } else {
+          debugPrint('Failed to load all groups for url {$url}. Response: $responseCode $responseBody');
+        }
       } catch (e) {
         debugPrint(e.toString());
       }
@@ -303,7 +335,7 @@ class Groups with Service implements NotificationsListener {
       return null;
     }
     String encodedTExt = Uri.encodeComponent(searchText);
-    String url = '${Config().groupsUrl}/groups?title=$encodedTExt';
+    String url = '${Config().groupsUrl}/v2/groups?title=$encodedTExt';
     Response? response = await Network().get(url, auth: Auth2());
     int responseCode = response?.statusCode ?? -1;
     String? responseBody = response?.body;
@@ -318,35 +350,17 @@ class Groups with Service implements NotificationsListener {
 
   Future<Group?> loadGroup(String? groupId) async {
     await waitForLogin();
-    if(StringUtils.isNotEmpty(groupId)) {
-      String url = '${Config().groupsUrl}/groups/$groupId';
+    if (StringUtils.isNotEmpty(groupId)) {
+      String url = '${Config().groupsUrl}/v2/groups/$groupId';
       try {
         Response? response = await Network().get(url, auth: Auth2(),);
         int responseCode = response?.statusCode ?? -1;
         String? responseBody = response?.body;
-        Map<String, dynamic>? groupsJson = ((responseBody != null) && (responseCode == 200)) ? JsonUtils.decodeMap(responseBody) : null;
-        return groupsJson != null ? Group.fromJson(groupsJson) : null;
-      } catch (e) {
-        debugPrint(e.toString());
-      }
-    }
-    return null;
-  }
-
-  //TBD sync with backend team, update group model and UI
-  Future<Group?> loadGroupByCanvasCourseId(int? courseId) async {
-    await waitForLogin();
-    if (courseId != null) {
-      String url = '${Config().groupsUrl}/groups/canvas_course/$courseId';
-      try {
-        Response? response = await Network().get(url, auth: Auth2());
-        int responseCode = response?.statusCode ?? -1;
-        String? responseBody = response?.body;
         if (responseCode == 200) {
-          Map<String, dynamic>? groupJson = JsonUtils.decodeMap(responseBody);
-          return Group.fromJson(groupJson);
+          Map<String, dynamic>? groupsJson = JsonUtils.decodeMap(responseBody);
+          return Group.fromJson(groupsJson);
         } else {
-          Log.d('Failed to load group by canvas course id. Reason: \n$responseCode: $responseBody');
+          debugPrint('Failed to load group with id {$groupId}. Response: $responseCode $responseBody');
         }
       } catch (e) {
         debugPrint(e.toString());
@@ -453,7 +467,94 @@ class Groups with Service implements NotificationsListener {
     }
   }
 
+  // Group Stats
+
+  Future<GroupStats?> loadGroupStats(String? groupId) async {
+    if (StringUtils.isEmpty(groupId)) {
+      return null;
+    }
+    await waitForLogin();
+    if (StringUtils.isNotEmpty(groupId)) {
+      String url = '${Config().groupsUrl}/group/$groupId/stats';
+      try {
+        Response? response = await Network().get(url, auth: Auth2());
+        int responseCode = response?.statusCode ?? -1;
+        String? responseBody = response?.body;
+        if (responseCode == 200) {
+          Map<String, dynamic>? statsJson = JsonUtils.decodeMap(responseBody);
+          return GroupStats.fromJson(statsJson);
+        } else {
+          debugPrint('Failed to load group stats for group {$groupId}. Reason: $responseCode, $responseBody');
+        }
+      } catch (e) {
+        debugPrint(e.toString());
+      }
+    }
+    return null;
+  }
+
   // Members APIs
+
+  Future<List<Member>?> loadMembers({String? groupId, List<GroupMemberStatus>? statuses, String? memberId, String? userId, 
+    String? externalId, String? netId, String? name, int? offset, int? limit}) async {
+    if (StringUtils.isEmpty(groupId)) {
+      debugPrint('Failed to load group members - missing groupId.');
+      return null;
+    }
+    await waitForLogin();
+    if (Config().groupsUrl != null) {
+      String url = '${Config().groupsUrl}/group/$groupId/members';
+      Map<String, dynamic> params = {};
+      if (CollectionUtils.isNotEmpty(statuses)) {
+        List<String> statusList = [];
+        for (GroupMemberStatus status in statuses!) {
+          statusList.add(groupMemberStatusToString(status)!);
+        }
+        params.addAll({'statuses': statusList});
+      }
+      if (StringUtils.isNotEmpty(memberId)) {
+        params.addAll({'id': memberId});
+      }
+      if (StringUtils.isNotEmpty(userId)) {
+        params.addAll({'user_id': userId});
+      }
+      if (StringUtils.isNotEmpty(externalId)) {
+        params.addAll({'external_id': externalId});
+      }
+      if (StringUtils.isNotEmpty(netId)) {
+        params.addAll({'net_id': netId});
+      }
+      if (StringUtils.isNotEmpty(name)) {
+        params.addAll({'name': name});
+      }
+      if (offset != null) {
+        params.addAll({'offset': offset});
+      }
+      if (limit != null) {
+        params.addAll({'limit': limit});
+      }
+      try {
+        String? body = JsonUtils.encode(params);
+        Response? response = await Network().get(url, auth: Auth2(), body: body);
+        int responseCode = response?.statusCode ?? -1;
+        String? responseBody = response?.body;
+        if (responseCode == 200) {
+          List<dynamic>? membersJson = JsonUtils.decodeList(responseBody);
+          return Member.listFromJson(membersJson);
+        } else {
+          debugPrint('Failed to load members for group $groupId with body {$body}. Reason: $responseCode $responseBody');
+        }
+      } catch (e) {
+        debugPrint(e.toString());
+      }
+    }
+    return null;
+  }
+
+  /// Admins and members are allowed to post.
+  Future<List<Member>?> loadMembersAllowedToPost({String? groupId}) async {
+    return await loadMembers(groupId: groupId, statuses: [GroupMemberStatus.admin, GroupMemberStatus.member]);
+  }
 
   Future<bool> requestMembership(Group? group, List<GroupMembershipAnswer>? answers) async{
     await waitForLogin();
@@ -538,12 +639,12 @@ class Groups with Service implements NotificationsListener {
     return false; // fail
   }
 
-  Future<bool> updateMembership(Group? group, Member? member, GroupMemberStatus status) async{
+  Future<bool> updateMembership(Group? group, String? memberId, GroupMemberStatus status) async{
     await waitForLogin();
-    if(StringUtils.isNotEmpty(group?.id) && StringUtils.isNotEmpty(member?.id)) {
+    if(StringUtils.isNotEmpty(group?.id) && StringUtils.isNotEmpty(memberId)) {
       Map<String, dynamic> bodyMap = {"status":groupMemberStatusToString(status)};
       String? body = JsonUtils.encode(bodyMap);
-      String url = '${Config().groupsUrl}/memberships/${member!.id}';
+      String url = '${Config().groupsUrl}/memberships/$memberId';
       try {
         Response? response = await Network().put(url, auth: Auth2(), body: body);
         if((response?.statusCode ?? -1) == 200){
@@ -564,10 +665,10 @@ class Groups with Service implements NotificationsListener {
     return false; // fail
   }
 
-  Future<bool> deleteMembership(Group? group, Member? member) async{
+  Future<bool> deleteMembership(Group? group, String? memberId) async{
     await waitForLogin();
-    if(StringUtils.isNotEmpty(group?.id) && StringUtils.isNotEmpty(member?.id)) {
-      String url = '${Config().groupsUrl}/memberships/${member!.id}';
+    if(StringUtils.isNotEmpty(group?.id) && StringUtils.isNotEmpty(memberId)) {
+      String url = '${Config().groupsUrl}/memberships/$memberId';
       try {
         Response? response = await Network().delete(url, auth: Auth2(),);
         if((response?.statusCode ?? -1) == 200){
@@ -1073,7 +1174,8 @@ class Groups with Service implements NotificationsListener {
 
   static Future<String?> _loadUserGroupsStringFromNet() async {
     if (StringUtils.isNotEmpty(Config().groupsUrl) && Auth2().isLoggedIn) {
-      Response? response = await Network().get('${Config().groupsUrl}/user/groups', auth: Auth2());
+      // Load all user groups because we cache them and use them for various checks on startup like flexUI etc
+      Response? response = await Network().get('${Config().groupsUrl}/v2/user/groups', auth: Auth2());
       if (response?.statusCode == 200) {
         return response?.body;
       }
@@ -1211,8 +1313,6 @@ class Groups with Service implements NotificationsListener {
   void _addAttendedMemberToCache({required Group group, required Member member}) {
     String groupId = group.id!;
     _attendedMembers ??= HashMap<String, List<Member>>();
-    group.members ??= <Member>[];
-    group.members!.add(member);
     List<Member>? attendedGroupMembers = _attendedMembers![groupId];
     attendedGroupMembers ??= <Member>[];
     attendedGroupMembers.add(member);
@@ -1260,42 +1360,6 @@ class Groups with Service implements NotificationsListener {
     }
   }
 
-  /// 
-  /// Initializes cached user groups with cached attended members if there are differences. 
-  /// So that user groups are in sync if the device is offline.
-  /// This is done only on startup
-  /// 
-  void _initUserGroupsWithCachedAttendedMembers() {
-    if (Connectivity().isOffline && CollectionUtils.isNotEmpty(_attendedMembers?.keys)) {
-      for (String groupId in _attendedMembers!.keys) {
-        Group? userGroup = _getUserGroup(groupId: groupId);
-        if (userGroup != null) {
-          List<Member>? attendedGroupMembers = _attendedMembers![groupId];
-          if (CollectionUtils.isNotEmpty(attendedGroupMembers)) {
-            List<Member>? groupMembers = userGroup.members;
-            if (CollectionUtils.isNotEmpty(groupMembers)) {
-              for (Member attendedMember in attendedGroupMembers!) {
-                Member? groupMember;
-                for (Member grMember in groupMembers!) {
-                  if (grMember.externalId == attendedMember.externalId) {
-                    groupMember = grMember;
-                    break;
-                  }
-                }
-                if (groupMember != null) {
-                  groupMember.dateAttendedUtc = attendedMember.dateAttendedUtc;
-                } else {
-                  userGroup.members ??= <Member>[];
-                  userGroup.members!.add(attendedMember);
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-
   Group? _getUserGroup({required String groupId}) {
     if (CollectionUtils.isNotEmpty(userGroups) && StringUtils.isNotEmpty(groupId)) {
       for (Group group in userGroups!) {
@@ -1309,10 +1373,16 @@ class Groups with Service implements NotificationsListener {
 
   // Report Abuse
 
-  Future<bool> reportAbuse({String? groupId, String? postId, String? comment}) async {
+  Future<bool> reportAbuse({String? groupId, String? postId, String? comment, bool reportToDeanOfStudents = false, bool reportToGroupAdmins = false}) async {
     if (Config().groupsUrl != null) {
-      Response? response = await Network().put('${Config().groupsUrl}/group/$groupId/posts/$postId/report/abuse',
-        body: JsonUtils.encode({'comment': comment}), auth: Auth2());
+      String url = '${Config().groupsUrl}/group/$groupId/posts/$postId/report/abuse';
+      String? body = JsonUtils.encode({
+        'comment': comment,
+        'send_to_dean_of_students': reportToDeanOfStudents,
+        'send_to_group_admins': reportToGroupAdmins,
+        
+      });
+      Response? response = await Network().put(url, body: body, auth: Auth2());
       return (response?.statusCode == 200);
     }
     return false;